Refactor file structure of the extension code
