+++ src/angular/index.ts
@@ -1,0 +1,2 @@
+export * from "./child";
+export * from "./parent";
