+++ src/extension.ts
@@ -1,20 +1,36 @@
-import { exec, ExecException, execSync } from "child_process";
-import * as fs from "fs";
-import * as path from "path";
+import { exec, ExecException } from "child_process";
 import * as vscode from "vscode";
-import { getConfig } from "./config";
-import { preRunChecks } from "./preRunChecks";
+import {
+  adjustParentComponentTemplate,
+  writeChildTemplate as writeChildComponentTemplate,
+} from "./angular";
+import { getComponentName } from "./utils/getComponentName";
+import { getDirectoryName } from "./utils/getDirectoryName";
+import { getExtensionId } from "./utils/getExtensionId";
+import { isAngularCliAvailable } from "./utils/isAngularCliAvailable";
+import { preRunChecks } from "./utils/preRunChecks";
 
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
 export const activate = (context: vscode.ExtensionContext) => {
-  // Use the console to output diagnostic information (console.log) and errors (console.error)
-  // This line of code will only be executed once when your extension is activated
   console.log(
     'Congratulations, your extension "angular-component-extractor" is now active!'
   );
 
-  const generateCommand = async (): Promise<void> => {
+  const generateCommand = getExtractCommand(context);
+  const disposable = vscode.commands.registerCommand(
+    "angular-component-extractor.extract-component",
+    generateCommand
+  );
+
+  context.subscriptions.push(disposable);
+};
+
+/**
+ * Get the command function to extract a child component
+ * @param context Context of the extension
+ * @returns Function which will be called when the command is executed
+ */
+const getExtractCommand = (context: vscode.ExtensionContext) => {
+  return async (): Promise<void> => {
     const editor = preRunChecks(
       vscode.window.activeTextEditor,
       vscode.extensions.getExtension,
@@ -52,7 +68,11 @@ export const activate = (context: vscode.ExtensionContext) => {
     );
 
     try {
-      writeNewComponent(componentDirectory, componentName, selectedSnippet);
+      writeChildComponentTemplate(
+        componentDirectory,
+        componentName,
+        selectedSnippet
+      );
     } catch (error) {
       vscode.window.showErrorMessage(
         `Could not write selected Code Snipped into generated component ${componentName}`
@@ -62,7 +82,7 @@ export const activate = (context: vscode.ExtensionContext) => {
 
     try {
       const extensionId = getExtensionId(context);
-      adjustCurrentComponent(componentName, extensionId);
+      adjustParentComponentTemplate(componentName, extensionId);
     } catch (error) {
       vscode.window.showErrorMessage(
         `Could not replace selected code with <${componentName}></${componentName}>`
@@ -70,62 +90,6 @@ export const activate = (context: vscode.ExtensionContext) => {
       return;
     }
   };
-
-  // The command has been defined in the package.json file
-  // Now provide the implementation of the command with registerCommand
-  // The commandId parameter must match the command field in package.json
-  const disposable = vscode.commands.registerCommand(
-    "angular-component-extractor.extract-component",
-    generateCommand
-  );
-
-  context.subscriptions.push(disposable);
-};
-
-/**
- * Prompt the user to enter the name of the component which will be created
- * @returns Promise of the component name
- */
-const getComponentName = (): Thenable<string | undefined> =>
-  vscode.window.showInputBox({
-    placeHolder: "Component name",
-    prompt: "Enter component name",
-    validateInput: (input: string) => {
-      // only lower or upper case letters and dashes allowed
-      const pattern = new RegExp(/^[a-z\-]+$/i);
-
-      return pattern.test(input)
-        ? undefined
-        : "Please enter a valid component name";
-    },
-  });
-
-/**
- * Check availability of Angular CLI on the command line
- * @param directory Directory which will use the Angular CLI
- */
-const isAngularCliAvailable = (directory: string): boolean => {
-  try {
-    execSync("ng --version", { cwd: directory });
-    return true;
-  } catch (error) {
-    console.error(error);
-    return false;
-  }
-};
-
-/**
- * Get the name of the directory of the current file
- * @returns Name of the directory
- */
-const getDirectoryName = (
-  document: vscode.TextDocument
-): string | undefined => {
-  const fileName = document.fileName;
-  if (!fileName) {
-    return undefined;
-  }
-  return path.dirname(fileName);
 };
 
 /**
@@ -177,55 +141,6 @@ const generateComponentProgress = (
       }
     );
   });
-};
-
-const writeNewComponent = (
-  directory: string,
-  componentName: string,
-  content: string
-) => {
-  const componentPath = path.join(
-    directory,
-    componentName,
-    `${componentName}.component.html`
-  );
-  const componentUri = vscode.Uri.file(componentPath);
-  vscode.workspace.fs.writeFile(componentUri, Buffer.from(content));
-};
-
-const adjustCurrentComponent = (componentName: string, extensionId: string) => {
-  // Get the active text editor
-  const editor = vscode.window.activeTextEditor;
-
-  const document = editor?.document;
-  const selection = editor?.selection;
-
-  if (!editor || !document || !selection) {
-    throw Error("Something went wrong");
-  }
-
-  // Get the default prefix from the VS Code settings
-  const defaultPrefix = getConfig(extensionId, "default-prefix");
-
-  // Component HTML tags
-  const component = `<${defaultPrefix}-${componentName}></${defaultPrefix}-${componentName}>`;
-  editor.edit((editBuilder) => {
-    editBuilder.replace(selection, component);
-  });
-};
-
-/**
- * Load the extension id of the package json file
- * @param context Context of VS Code extension
- */
-const getExtensionId = (context: vscode.ExtensionContext): string => {
-  const extensionPath = path.join(context.extensionPath, "package.json");
-  const packageFile = JSON.parse(fs.readFileSync(extensionPath, "utf8"));
-
-  if (!packageFile) {
-    throw Error("Failed to initialize extension!");
-  }
-  return packageFile.name;
 };
 
 // this method is called when your extension is deactivated
